Move CLI related commands to `nu-cli`
Part of the larger cratification effort.

Moves all `reedline` or shell line editor specific commands to `nu-cli`

From `nu-cmd-lang`:
- `commandline`
  - This shouldn't have moved there. Doesn't directly depend on reedline
but assumes parts in the engine state that are specific to the use of
reedline or a REPL

From `nu-command`:
- `keybindings` and subcommands
  - `keybindings default`
  - `keybindings list`
  - `keybindings listen`
    - very `reedline` specific
- `history`
  - needs `reedline`
- `history session`

Instead of having a separate `create_default_context()` that calls
`nu-command`'s `create_default_context()` I added a `add_cli_context()`
that updates an `EngineState`

I removed the `test_example()` facilities for now as we did not run any
of the commands in an `Example` test and importing the right context for
those tests seemed more of a hassle than the duplicated
`test_examples()` implementations in `nu-cmd-lang` and `nu-command`
