+++ crates/nu-cli/Cargo.toml
@@ -38,6 +38,7 @@ miette = { version = "5.7.0", features = ["fancy-no-backtrace"] }
 percent-encoding = "2"
 sysinfo = "0.28.2"
 thiserror = "1.0.31"
+unicode-segmentation = "1.10.0"
 
 [features]
 plugin = []
