+++ crates/nu-cli/src/commands/commandline.rs
