+++ crates/nu-cli/src/commands/history.rs
