+++ crates/nu-cli/src/commands/history_session.rs
