+++ crates/nu-cli/src/commands/keybindings.rs
