+++ crates/nu-cli/src/commands/keybindings_default.rs
