+++ crates/nu-cli/src/commands/keybindings_list.rs
