+++ crates/nu-cli/src/commands/mod.rs
@@ -1,9 +1,18 @@
+mod commandline;
+mod default_context;
+mod history;
+mod history_session;
 mod keybindings;
 mod keybindings_default;
 mod keybindings_list;
 mod keybindings_listen;
 
+pub use commandline::Commandline;
+pub use history::History;
+pub use history_session::HistorySession;
 pub use keybindings::Keybindings;
 pub use keybindings_default::KeybindingsDefault;
 pub use keybindings_list::KeybindingsList;
 pub use keybindings_listen::KeybindingsListen;
+
+pub use default_context::add_cli_context;
