+++ crates/nu-cli/src/eval_cmds.rs
