+++ crates/nu-cmd-lang/Cargo.toml
@@ -24,7 +24,6 @@ fancy-regex = "0.11.0"
 itertools = "0.10.0"
 log = "0.4.14"
 shadow-rs = { version = "0.21.0", default-features = false }
-unicode-segmentation = "1.10.0"
 
 [build-dependencies]
 shadow-rs = { version = "0.21.0", default-features = false }
