+++ crates/nu-cmd-lang/src/core_commands/mod.rs
@@ -1,7 +1,6 @@
 mod alias;
 mod break_;
 mod collect;
-mod commandline;
 mod const_;
 mod continue_;
 mod def;
@@ -43,7 +42,6 @@ mod while_;
 pub use alias::Alias;
 pub use break_::Break;
 pub use collect::Collect;
-pub use commandline::Commandline;
 pub use const_::Const;
 pub use continue_::Continue;
 pub use def::Def;
