+++ crates/nu-command/src/misc/mod.rs
@@ -1,7 +1,3 @@
-mod history;
-mod history_session;
 mod tutor;
 
-pub use history::History;
-pub use history_session::HistorySession;
 pub use tutor::Tutor;
