+++ crates/nu-command/src/platform/mod.rs
@@ -4,7 +4,6 @@ mod dir_info;
 mod du;
 mod input;
 mod kill;
-mod reedline_commands;
 mod sleep;
 mod term_size;
 
@@ -14,6 +13,5 @@ pub use dir_info::{DirBuilder, DirInfo, FileInfo};
 pub use du::Du;
 pub use input::Input;
 pub use kill::Kill;
-pub use reedline_commands::{Keybindings, KeybindingsDefault, KeybindingsList, KeybindingsListen};
 pub use sleep::Sleep;
 pub use term_size::TermSize;
