+++ src/main.rs
@@ -42,7 +42,7 @@ fn main() -> Result<()> {
 
     // Get initial current working directory.
     let init_cwd = get_init_cwd();
-    let mut engine_state = create_default_context();
+    let mut engine_state = nu_cli::add_cli_context(create_default_context());
 
     // Custom additions
     let delta = {
